requirements.txt
beautifulsoup4==4.12.3
pydantic==2.6.4
python-dotenv==1.0.1
python_dateutil==2.9.0.post0
Requests==2.31.0
